Add editable category and integrations tests.
